requirements.txt
Django==5.1.3

# Main dependencies
django-tables2
django_polymorphic
django-select2
django-filter

# Visual dependencies
crispy-bootstrap5
django-crispy-forms
markdown
PyPDF2

# external connections
boto3

# analysis 
pandas

